+++ Code/lib/BWC_unified/FW_VERSION.h
@@ -1,1 +1,1 @@
-#define FW_VERSION "4.0.2"
+#define FW_VERSION "4.0.3"